requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
Pillow==10.2.0
streamlit==1.31.0

